Move test into tests/ dir
